Fixes, improvements, refractor
