+++ settings.cfg
@@ -8,9 +8,12 @@ projectURL = https://github.com/Ethck/Path-of-Accounting/
 releaseURL = https://api.github.com/repos/Ethck/Path-of-Accounting/releases
 version = v0.8
 
+# If we get limited results the program will try to remove mods listed here
+badMods = ["Reflects # Physical Damage to Melee Attackers", "FAKE MOD"]
+
 [GUI]
-useGUI = no
-timeout = 3
+useGUI = yes
+timeout = 10
 
 # To change colors go to https://www.google.com/search?q=color+picker and copy/paste the HEX code for the colors you want
 backgroundColor = #1a1a1a
